# docs/swagger/swagger.yaml
definitions:
...
  dto.SubscribeRequest:
    properties:
      billing_cycle:
        enum:
        - monthly
        - yearly
        type: string
      package_code:
        type: string
      payment_method:
        type: string
    required:
    - billing_cycle
    - package_code
    - payment_method
    type: object
  dto.UnlockContactRequest:
    properties:
      idempotency_key:
        type: string
      product_id:
        type: string
      seller_id:
        type: string
    required:
    - product_id
    - seller_id
    type: object
  dto.UpgradeSubscriptionRequest:
    properties:
      billing_cycle:
        enum:
        - monthly
        - yearly
        type: string
      package_code:
        type: string
      payment_method:
        type: string
    required:
    - billing_cycle
    - package_code
    - payment_method
    type: object
info:
  contact: {}
  description: Payment service API (Gin)
  title: Modami Payment Service
  version: "1.0"
paths:
...
  /packages:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped []dto.PackageResponse
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List available packages
      tags:
      - Packages
  /packages/{code}:
    get:
      parameters:
      - description: Package code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.PackageResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get package by code
      tags:
      - Packages
...
  /subscriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Subscribe request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscribeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Wrapped dto.SubscriptionResponse (with payment_url)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe to a package (creates pending subscription + payment URL)
      tags:
      - Subscriptions
  /subscriptions/current:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.SubscriptionResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current subscription for current user
      tags:
      - Subscriptions
  /subscriptions/current/cancel:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.CancelSubscriptionResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Disable auto-renew for current subscription
      tags:
      - Subscriptions
  /subscriptions/current/renew:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Re-enable auto-renew for current subscription
      tags:
      - Subscriptions
  /subscriptions/history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped []dto.SubscriptionResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List subscription history for current user
      tags:
      - Subscriptions
  /subscriptions/upgrade:
    post:
      consumes:
      - application/json
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Upgrade request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpgradeSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Wrapped dto.SubscriptionResponse (with payment_url)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upgrade subscription (creates new pending subscription + payment URL)
      tags:
      - Subscriptions
  /unlocks:
    get:
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.UnlockListResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List unlocks for current user
      tags:
      - Unlocks
    post:
      consumes:
      - application/json
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Unlock request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UnlockContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Wrapped dto.UnlockContactResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unlock a product contact (deduct credits)
      tags:
      - Unlocks
  /unlocks/check/{product_id}:
    get:
      parameters:
      - description: Product ID (uuid)
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.CheckUnlockResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check if product already unlocked
      tags:
      - Unlocks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"